requirements.txt
Django==3.1.7


django-filter==2.4.0